# runtime/expr: fix subbenchmark names in BenchmarkSort