--- a/runtime/expr/sort_test.go
+++ b/runtime/expr/sort_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/brimdata/zed"
+	"github.com/brimdata/zed/zson"
 )
 
 func BenchmarkSort(b *testing.B) {
@@ -18,7 +19,7 @@ func BenchmarkSort(b *testing.B) {
 		{zed.TypeString, func() []byte { return strconv.AppendUint(nil, rand.Uint64(), 16) }},
 	}
 	for _, c := range cases {
-		b.Run(c.typ.Kind().String(), func(b *testing.B) {
+		b.Run(zson.FormatType(c.typ), func(b *testing.B) {
 			cmp := NewComparator(false, false, &This{})
 			vals := make([]zed.Value, 1048576)
 			var sorter Sorter