Before:

    $ go test ./runtime/expr -bench BenchmarkSort
    goos: darwin
    goarch: arm64
    pkg: github.com/brimdata/zed/runtime/expr
    BenchmarkSort/primitive-10         	       3	 473237667 ns/op
    BenchmarkSort/primitive#01-10      	       1	1019121708 ns/op
    BenchmarkSort/primitive#02-10      	       1	1435500792 ns/op
    PASS
    ok  	github.com/brimdata/zed/runtime/expr	6.188s

After:

    $ go test ./runtime/expr -bench BenchmarkSort
    goos: darwin
    goarch: arm64
    pkg: github.com/brimdata/zed/runtime/expr
    BenchmarkSort/int64-10         	       3	 471964417 ns/op
    BenchmarkSort/uint64-10        	       1	1009214541 ns/op
    BenchmarkSort/string-10        	       1	1433023666 ns/op
    PASS
    ok  	github.com/brimdata/zed/runtime/expr	6.169s